+++ lib/cnext/include/Coroutines.h
@@ -341,7 +341,7 @@ msg_t* comessageQueuePeek(void);
 msg_t* comessageQueuePop(void);
 msg_t* comessageQueuePopType(int type);
 msg_t* comessageQueueWait(const struct timespec *ts);
-msg_t* comessageQueueWaitForType(int type, const struct timespec *ts);
+msg_t* comessageQueueWaitForType(int64_t type, const struct timespec *ts);
 int comessageQueuePush(Coroutine *coroutine, msg_t *comessage);
